Refactor: Decouple hint logic into strategy-based architecture
This commit refactors the hint generation system by moving specific hint-finding logic from `HintProvider` into dedicated strategy classes. This improves modularity, testability, and adheres to the strategy design pattern.

**Key Changes:**

*   **Strategy Pattern Implementation**:
    *   Introduced the `HintStrategy` interface.
    *   Created specific implementations: `NakedSingleStrategy`, `HiddenSingleStrategy`, `PointingCandidateStrategy`, and `ClaimingCandidateStrategy`.
    *   `HintProvider` now coordinates these strategies rather than containing the implementation logic.

*   **Refactor `HintProvider`**:
    *   Updated the constructor to accept strategy implementations (supporting dependency injection).
    *   Simplified `provideHint` to delegate to strategy classes.
    *   Removed internal helper methods now handled by strategies or utility functions.

*   **Code Organization**:
    *   **`Hint.kt`**: Relocated `Hint`, `HintType`, and `GroupType` definitions to a dedicated file.
    *   **`HintUtils.kt`**: Extracted common grid navigation helpers like `getRow`, `getColumn`, `getBlock`, and `containsCell` into a separate utility file.
    *   **Dependency Injection**: Updated `Module.kt` to register the new strategies and the updated `HintProvider`.

*   **Testing**:
    *   Updated `HintProviderTest` to reflect the new architecture, utilizing individual strategies for specific hint validation while maintaining coverage for the provider.
