+++ sudoku-core/src/main/java/io/github/mgrablo/sudokucore/hints/strategies/HintStrategy.kt
@@ -1,0 +1,9 @@
+package io.github.mgrablo.sudokucore.hints.strategies
+
+import io.github.mgrablo.sudokucore.hints.Hint
+import io.github.mgrablo.sudokucore.model.House
+import io.github.mgrablo.sudokucore.model.SudokuCellData
+
+internal interface HintStrategy {
+	fun findHints(data: List<SudokuCellData>, houses: List<House>): List<Hint>
+}
